third_party/WebKit/LayoutTests/platform/mac-mac10.11/virtual/threaded/fast/scroll-behavior/overscroll-behavior-expected.txt

This is a testharness.js-based test.
PASS overscroll-behavior-y: none should only prevent scroll propagation on y axis with: wheel.
FAIL overscroll-behavior-x: none should only prevent scroll propagation on x axis with: wheel. assert_equals: expected 100 but got 66
PASS overscroll-behavior should not affect scrolling inside the applied container with: wheel.
PASS overscroll-behavior on non-scrollable area should not affect scroll propagation with: wheel.
Harness: the test ran to completion.
